requirements.txt
streamlit
faker
durable-rules
